requirements.txt
google-cloud-asset
pandas
google-cloud-bigquery
